refactor: repository 패키지명 query로 변경 (#105)
query에 대한 구현체 (Kotlin JDSL)는 이 패키지에 생성 예정
